+++ app/src/main/java/com/cmt/openapp/detail/data/DetailRepository.kt
@@ -6,7 +6,7 @@ import retrofit2.Response
 import javax.inject.Inject
 
 class DetailRepository @Inject constructor(private val detailClient: DetailClient) {
-    suspend fun getIncidentDetail(id: Long): Response<IncidentDTODetail> {
+    suspend fun getIncidentDetail(id: Int): Response<IncidentDTODetail> {
         return detailClient.getIncidentDetail(id)
     }
 }
