+++ app/src/main/java/com/cmt/openapp/detail/ui/viewmodel/DetailViewModel.kt
@@ -22,7 +22,7 @@ class DetailViewModel @Inject constructor(private val repository: DetailReposito
     private val _isLoading = MutableStateFlow(true)
     val isLoading: StateFlow<Boolean> = _isLoading
 
-    fun loadIncidentDetail(id: Long) {
+    fun loadIncidentDetail(id: Int) {
         _isLoading.value = true
         viewModelScope.launch(Dispatchers.IO) {
             runCatching {
